model_image_ocr/requirements.txt
# OCR Toolkit 依赖

# 核心依赖
paddlepaddle-gpu>=2.5.0  # 或 paddlepaddle (CPU版本)
paddleocr>=2.7.0

# 图像处理
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0

# GUI
gradio>=4.0.0

# 数据处理
pandas>=1.3.0

# ONNX导出 (可选)
paddle2onnx>=1.0.0
onnx>=1.12.0
onnxruntime-gpu>=1.12.0  # 或 onnxruntime (CPU版本)

# TensorRT部署 (可选，需要单独安装)
# tensorrt>=8.0
# pycuda>=2022.1
